+++ src/main/java/com/pos/dto/response/OrderItemResponse.java
@@ -13,7 +13,7 @@ public class OrderItemResponse {
     private Long productId;
     private String productName;
     private String productSku;
-    private int quantity;
+    private BigDecimal quantity;
     private BigDecimal unitPrice;
     private BigDecimal subtotal;
